extracted AbstractPageFilterInputContainerComponentProcessor
